feat: agregar componente Chatbot y Map Generator, actualizar rutas en el módulo de IA
